requirement.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
pandas==1.2.3
Pillow==8.1.0
pygame==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
subprocess.run==0.0.8
